requirements.txt
streamlit
faiss-cpu
sentence-transformers
python-docx
python-pptx
pdfplumber
pandas
openai
